first draft missingness viz
I added some NA values to the iris data set, and it worked using the simpler version of melt.  More complicated tables will benefit from the better controlled version of melt.  Still needs set methods and a call to the actual visualization -- is the pdf creation code in upsetDB() reusable?